+++ driver-core/src/main/com/mongodb/internal/observability/micrometer/Span.java
@@ -56,6 +56,14 @@ public interface Span {
         }
 
         @Override
+        public void tagHighCardinality(final KeyValue keyValue) {
+        }
+
+        @Override
+        public void tagHighCardinality(final KeyValues keyValues) {
+        }
+
+        @Override
         public void tagHighCardinality(final String keyName, final BsonDocument value) {
         }
 
@@ -96,6 +104,20 @@ public interface Span {
      * @param keyValues The key-value pairs representing the tags.
      */
     void tagLowCardinality(KeyValues keyValues);
+
+    /**
+     * Adds a high-cardinality tag to the span.
+     *
+     * @param keyValue The key-value pair representing the tag.
+     */
+    void tagHighCardinality(KeyValue keyValue);
+
+    /**
+     * Adds multiple high-cardinality tags to the span.
+     *
+     * @param keyValues The key-value pairs representing the tags.
+     */
+    void tagHighCardinality(KeyValues keyValues);
 
     /**
      * Adds a high-cardinality (highly variable values) tag to the span with a BSON document value.
